refactor(gamma): move MemberAvatar into shared/components/
Consolidates the groups feature's duplicate of MemberAvatar (byte-for-byte
identical to the one already used across features/applications/user-permissions)
into shared/components/, and updates GroupAddMembersSheet/GroupRemoveMemberSheet
to import it from there.
