+++ gravitee-gamma/gravitee-gamma-module-platform/src/main/ui/shared/components/MemberAvatar.tsx
